agents/email/requirements.txt
python-dotenv
html2text
imapclient